backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic>=2.0.0
pymysql==1.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings>=2.0.0
python-multipart==0.0.9
